fix: Stop the calling of style.css on every reload
Why this was happening?
express.static(path) middleware redirects the request
to static resources from *root* (ex: `/style.css`) to
the path.

But since, in the views, I was requesting for `style.css`
instead of `/style.css` for the style resources, this caused
the browser to request for `style.css` from the current folder

Hencewhy causing the error of matching into the categoryId that
was seen. Easily fixed by just changing the path.